+++ src/common/chains.ts
@@ -1,0 +1,5 @@
+export enum SupportedChainId {
+  MAINNET = 1,
+  GOERLI = 5,
+  GNOSIS_CHAIN = 100,
+}
