+++ src/common/ipfs.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_IPFS_READ_URI = 'https://gnosis.mypinata.cloud/ipfs'
+export const DEFAULT_IPFS_WRITE_URI = 'https://api.pinata.cloud'
