--- src/constants/chains.ts
@@ -1,11 +1,0 @@
-export enum SupportedChainId {
-  MAINNET = 1,
-  GOERLI = 5,
-  GNOSIS_CHAIN = 100,
-}
-
-export const ALL_SUPPORTED_CHAIN_IDS: SupportedChainId[] = [
-  SupportedChainId.MAINNET,
-  SupportedChainId.GOERLI,
-  SupportedChainId.GNOSIS_CHAIN,
-]
