--- src/types/index.ts
@@ -1,9 +1,0 @@
-export * from '../api/order-book/types'
-export * from '../api/metadata/types'
-export * from './sdk'
-export * from './utilities'
-
-export { OrderKind } from '@cowprotocol/contracts'
-export class Token {
-  constructor(public symbol: string, public address: string) {}
-}
