--- src/types/sdk.ts
@@ -1,5 +1,0 @@
-import { LogLevelDesc } from 'loglevel'
-
-export type SdkOptions = {
-  loglevel?: LogLevelDesc
-}
